requirements.txt
# CORE DO DJANGO
Django==5.2.8
    # asgiref==3.11.0
    # sqlparse==0.5.3
    # typing_extensions==4.15.0
# Adaptador para PostgreSQL
psycopg2-binary==2.9.11
# para manipulação de imagens (thumbnails/previews)
pillow==12.0.0


# GERENCIAMENTO DE CONFIGURAÇÃO
# Para carregar settings de variáveis de ambiente
django-environ==0.12.0
django-cleanup==9.0.0


# ARMAZENAMENTO EM S3

django-storages==1.14.6
boto3==1.41.1
    # botocore==1.41.1
    # jmespath==1.0.1
    # python-dateutil==2.9.0.post0
    # s3transfer==0.15.0
    # six==1.17.0
    # urllib3==2.5.0


# DEPLOY E SERVIR ESTATICOS
whitenoise==6.11.0
